added hp, fps 1line, restart menu

+++ Class/Menu/Button.cs
@@ -32,10 +32,10 @@ namespace Menu
         public void Update()
         {
             if (Game1.Game1._mouseState.LeftButton == ButtonState.Pressed &&
-                Game1.Game1._mouseState.X < (int)_position.X + (int)_font.MeasureString(_text).X &&
-                Game1.Game1._mouseState.X > (int)_position.X &&
-                Game1.Game1._mouseState.Y < (int)_position.Y + (int)_font.MeasureString(_text).Y &&
-                Game1.Game1._mouseState.Y > (int)_position.Y &&
+                Game1.Game1._mouseState.X < _position.X + _font.MeasureString(_text).X &&
+                Game1.Game1._mouseState.X > _position.X &&
+                Game1.Game1._mouseState.Y < _position.Y + _font.MeasureString(_text).Y &&
+                Game1.Game1._mouseState.Y > _position.Y &&
                 !Game1.Game1.isLeftMouseButtonPressed
                 )
             {
